file_consumer/requirements.txt
kafka-python
cassandra-driver
Pygments==2.11.2